express routes added for refactoring
